Upgrade to expat 2.1.1
Modified the import_expat.sh script as the archive is now
compressed using bzip2 instead of gzip and uses configure.ac
instead of configure.in.

Modified README.android to make the example command up to date
with respect to the new archive format and also to ensure that
it builds dependencies.

All the remaining changes are straight from upstream.

Bug: 27818751
Change-Id: Ia4d587e98a3f68b7133cdbcb0ca952e18fffcfe8