Stop building libexpat with -fexceptions.
This flag is usually unnecessary when building C code. It is only useful
when throwing exceptions across C stack frames, which isn't something that
we generally support.

Passing -fexceptions can create a dependency on the unwinder when building
with HWASAN and possibly in other build configurations. With the change
to move the unwinder into a shared library:
https://android-review.googlesource.com/q/topic:%22libunwind-so%22
this would otherwise create an unnecessary dependency on the unwinder
shared object from libexpat.so.

We have been building expat with -fexceptions since the initial open source
release. It doesn't seem to have been added for any Android-specific
reason, but appears to have been inherited from the upstream build
system, which originally started passing it in this commit from 2001:
https://github.com/libexpat/libexpat/commit/d8263250ee5fe2eefc7cd94a05c5ee077912bc55

As far as I can tell, none of the code in the platform depends on expat
being built with -fexceptions. Very little code in the platform builds with
-fexceptions anyway, most notably a few libraries in external. One of these
libraries, skia, uses expat, but does not use exceptions in the translation
unit that uses expat. Chromium has the same dependency (skia -> expat)
and isn't building expat with -fexceptions either.

Bug: 144430859
Change-Id: I2188b30fb155e0e0241634fda6e45c6fe06c297a